requirements.txt
# CPU-only PyTorch (must be installed before sentence-transformers)
--extra-index-url https://download.pytorch.org/whl/cpu
torch

fastapi
uvicorn
gitpython
requests
python-dotenv
pydantic
aiofiles
tqdm
numpy
scipy
python-Levenshtein
faiss-cpu
sentence-transformers
scikit-learn
chardet
langgraph
gitpython
pandas
openpyxl
python-dotenv
google-generativeai
matplotlib
radon          
chardet
requests
supabase
postgrest
pytest
pytest-asyncio
pytest-cov
pytest-mock
httpx
faker
python-multipartredis
